+++ src/Bore/FrontMatter.hs
@@ -2,11 +2,14 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE RecordWildCards #-}
 
 module Bore.FrontMatter 
     ( FrontMatter(..)
     , isGophermap
     , parseFrontMatter
+    , updateFrontMatter
+    , splitFrontmatter
     ) where
 
 import GHC.Generics (Generic)
@@ -20,10 +23,10 @@ import Control.Exception (toException)
 import qualified Data.Text.Encoding as TE
 import qualified Data.Map as Map
 import Data.Data (Data, Typeable)
+import Control.Applicative ((<|>))
 
 data FrontMatter = FrontMatter {
     title :: Maybe Text.Text,
-    phlog :: Maybe Bool,
     date :: Maybe Text.Text,
     tags :: Maybe [Text.Text],
     lambdas :: Maybe [Text.Text],
@@ -37,6 +40,22 @@ data FrontMatter = FrontMatter {
     -- defined parent.
     variables :: Maybe (Map.Map Text.Text Text.Text)
     } deriving (Show, Data, Typeable, Generic, FromJSON)
+
+-- FIXME: I hate this.
+-- | A function to merge two FrontMatter records, with the second one taking precedence.
+updateFrontMatter :: FrontMatter -> FrontMatter -> FrontMatter
+updateFrontMatter base FrontMatter{..} = base
+  { title = base.title <|> title
+  , date = base.date <|> date
+  , tags = base.tags <|> tags
+  , lambdas = base.lambdas <|> lambdas
+  , draft = base.draft <|> draft
+  , skipMarkdown = base.skipMarkdown <|> skipMarkdown
+  , skipTemplating = base.skipTemplating <|> skipTemplating
+  , gophermap = base.gophermap <|> gophermap
+  , parent = base.parent <|> parent
+  , variables = base.variables <|> variables
+  }
 
 -- Automatically derive ToJSON for FrontMatter
 instance ToJSON FrontMatter
@@ -58,6 +77,7 @@ parseFrontMatter content =
     Just (frontmatterText, mainText) -> (, mainText) <$> decodeEither' (TE.encodeUtf8 frontmatterText)
     Nothing -> Left $ OtherParseException $ toException $ userError "No frontmatter found"
 
+-- FIXME: should only split if the document begins with ---\n, while currently this is too permissive.
 -- | Return the frontmatter and the rest of the text as separate Text values.
 --
 -- Currently is maybe too permissive in that in violation of the spec, I think,